[CDRB030_BurnIn(v1.1)]         1.add media play (raw/v_0189.mp3)         2.add command Receiver (start/close)         3.update to camera2 api (SurfaceTexture)         4.record video to sd card         5.1080p 27fps success (no audio)         6.config: mercury/storage/DCIM/BurnInTestConfig.ini         7.log: mercury/storage/DCIM/BurnInTestLog.ini

+++ app/src/main/java/com/d160/view/mLogMsg.java
@@ -5,7 +5,7 @@ import com.d160.wa034.Utils;
 import java.util.*;
 
 public class mLogMsg {
-    public int run;
+    public long run;
     public Date date;
     public String msg;
     public mLog type;
